backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pydantic
fuzzywuzzy
python-Levenshtein
redis
aiosmtplib
